DRF/day35/blog/requirements.txt
Django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt
psycopg2-binary>=2.9  
gunicorn>=21.2
django-cors-headers>=4.0
whitenoise
pillow
